+++ imod/tests/test_mf6/test_import_prj.py
@@ -174,34 +174,34 @@ def test_import_ipf(tmp_path):
     result_snippet_1 = open_projectfile_data(projects_file)
 
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L3"]["dataframe"][0]["rate"]
-        == 2 * result_snippet_0[0]["wel-WELLS_L3"]["dataframe"][0]["rate"] + 1.3
+        result_snippet_1[0]["wel-1-WELLS_L3"]["dataframe"][0]["rate"]
+        == 2 * result_snippet_0[0]["wel-1-WELLS_L3"]["dataframe"][0]["rate"] + 1.3
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L4"]["dataframe"][0]["rate"]
-        == -1 * result_snippet_0[0]["wel-WELLS_L4"]["dataframe"][0]["rate"] + 0
+        result_snippet_1[0]["wel-2-WELLS_L4"]["dataframe"][0]["rate"]
+        == -1 * result_snippet_0[0]["wel-2-WELLS_L4"]["dataframe"][0]["rate"] + 0
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L5"]["dataframe"][0]["rate"]
-        == 2 * result_snippet_0[0]["wel-WELLS_L4"]["dataframe"][0]["rate"] + 1.3
+        result_snippet_1[0]["wel-3-WELLS_L5"]["dataframe"][0]["rate"]
+        == 2 * result_snippet_0[0]["wel-3-WELLS_L5"]["dataframe"][0]["rate"] + 1.3
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L3"]["dataframe"][0]["filt_top"] == 11.0
+        result_snippet_1[0]["wel-1-WELLS_L3"]["dataframe"][0]["filt_top"] == 11.0
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L3"]["dataframe"][0]["filt_bot"] == 6.0
+        result_snippet_1[0]["wel-1-WELLS_L3"]["dataframe"][0]["filt_bot"] == 6.0
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L4"]["dataframe"][0]["filt_top"] == 11.0
+        result_snippet_1[0]["wel-2-WELLS_L4"]["dataframe"][0]["filt_top"] == 11.0
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L4"]["dataframe"][0]["filt_bot"] == 6.0
+        result_snippet_1[0]["wel-2-WELLS_L4"]["dataframe"][0]["filt_bot"] == 6.0
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L5"]["dataframe"][0]["filt_top"] == 11.0
+        result_snippet_1[0]["wel-3-WELLS_L5"]["dataframe"][0]["filt_top"] == 11.0
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L5"]["dataframe"][0]["filt_bot"] == 6.0
+        result_snippet_1[0]["wel-3-WELLS_L5"]["dataframe"][0]["filt_bot"] == 6.0
     )
 
 
@@ -243,13 +243,13 @@ def test_import_ipf_unique_id_and_logging(tmp_path):
     # test that id's were made unique
     # Assert
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L3"]["dataframe"][0]["id"] == "extractions"
+        result_snippet_1[0]["wel-1-WELLS_L3"]["dataframe"][0]["id"] == "extractions"
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L4"]["dataframe"][0]["id"] == "extractions_1"
+        result_snippet_1[0]["wel-2-WELLS_L4"]["dataframe"][0]["id"] == "extractions_1"
     )
     assert np.all(
-        result_snippet_1[0]["wel-WELLS_L5"]["dataframe"][0]["id"] == "extractions_2"
+        result_snippet_1[0]["wel-3-WELLS_L5"]["dataframe"][0]["id"] == "extractions_2"
     )
 
     with open(logfile_path, "r") as log_file: